feat: add PEP-561 py.typed file for downstream type checkers
This follows PEP-561 to mark syrupy as compatible with type
checking (driven by the type annotations that already exist and are
checked), by creating an empty `py.typed` in the project. This stops
users from having to ignore syrupy when type checking, and allows the
type checker to provide specific guidance about possible problems.

Fixes #439