refactor: simplify data serializer for ambr (#676)
* refactor: simplify data serializer for ambr

* feat: introduce concept of a tainted snapshot

BREAKING CHANGE: Serializers may now throw a TaintedSnapshotError which will tell the user to regenerate the snapshot even if the underlying data has not changed. This is to support rolling out more subtle changes to the serializers, such as the introduction of serializer metadata.

BREAKING CHANGE: Renamed DataSerializer to AmberDataSerializer.